+++ scrapper/src/main/java/edu/java/configuration/SchedulerConfig.java
@@ -1,0 +1,18 @@
+package edu.java.configuration;
+
+import edu.java.scheduler.LinkUpdaterScheduler;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class SchedulerConfig {
+
+    @Bean
+    @ConditionalOnProperty(name = "app.scheduler.enabled", matchIfMissing = true)
+    public LinkUpdaterScheduler linkUpdaterScheduler() {
+        return new LinkUpdaterScheduler();
+    }
+}
