+++ scrapper/src/main/java/edu/java/scheduler/LinkUpdaterScheduler.java
@@ -1,0 +1,14 @@
+package edu.java.scheduler;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
+
+@Slf4j
+public class LinkUpdaterScheduler {
+
+    @Scheduled(fixedDelayString = "#{@scheduler.interval.toMillis()}")
+    void update() {
+        log.info("тут будет проверка ссылок на обновление");
+    }
+
+}
